+++ pyriemann/utils/mean.py
@@ -481,9 +481,15 @@ def mean_power(X, p, *, sample_weight=None, zeta=10e-10, maxiter=100,
 
     if p == 1:
         return mean_euclid(X, sample_weight=sample_weight)
-    elif p == 0:
-        return mean_riemann(X, sample_weight=sample_weight)
-    elif p == -1:
+    if p == 0:
+        return mean_riemann(
+                X,
+                sample_weight=sample_weight,
+                init=init,
+                tol=zeta,
+                maxiter=maxiter,
+               )
+    if p == -1:
         return mean_harmonic(X, sample_weight=sample_weight)
 
     n_matrices, n, _ = X.shape