+++ doc/whatsnew.rst
@@ -17,6 +17,8 @@ v0.9.dev
 
 - Fix matplotlib warning. :pr:`351` by :user:`qbarthelemy`
 
+- Enhance :func:`pyriemann.utils.mean.mean_power` using ``init``, ``tol`` and ``maxiter`` parameters when p=0. :pr:`353` by :user:`toncho11`
+
 v0.8 (February 2025)
 --------------------